+++ doc/changelog.rst
@@ -28,6 +28,7 @@ Addons
 Front
 ~~~~~
 
+- Fix category view pagination
 - Fix category view rendering for ajax requests
 - Fix product search to only show searchable products
 - Rename `get_visible_products` to `get_listed_products`